+++ tests/src/rules/no-unresolved.js
@@ -343,9 +343,9 @@ ruleTester.run('no-unresolved unknown resolver', rule, {
     // logs resolver load error
     test({
       code: 'import "./malformed.js"',
-      settings: { 'import/resolver': 'foo' },
+      settings: { 'import/resolver': 'doesnt-exist' },
       errors: [
-        `Resolve error: unable to load resolver "foo".`,
+        `Resolve error: unable to load resolver "doesnt-exist".`,
         `Unable to resolve path to module './malformed.js'.`,
       ],
     }),
@@ -353,9 +353,9 @@ ruleTester.run('no-unresolved unknown resolver', rule, {
     // only logs resolver message once
     test({
       code: 'import "./malformed.js"; import "./fake.js"',
-      settings: { 'import/resolver': 'foo' },
+      settings: { 'import/resolver': 'doesnt-exist' },
       errors: [
-        `Resolve error: unable to load resolver "foo".`,
+        `Resolve error: unable to load resolver "doesnt-exist".`,
         `Unable to resolve path to module './malformed.js'.`,
         `Unable to resolve path to module './fake.js'.`,
       ],